+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - added optional caption bar
 - added new "large" popup template 
 - added a button to empty a layer without deleting it
+- added a button to clone a datalayer
 
 ## 0.5.x
 - datalayers are now sent to backend as geojson